refactor nodes hub page
