+++ utk_curio/frontend/urban-workflows/src/pages/nodes/my-packs/MyPacks.module.css
@@ -1,0 +1,285 @@
+/*
+ * Shared CSS module for the My Packs sidebar feature cluster.
+ *
+ * Used by:
+ *   MyPackIconActions.tsx
+ *   MyPackSingletonRow.tsx
+ *   HubInstalledForkRailGroup.tsx
+ *   MyPacksSidebar.tsx
+ */
+
+/* ── Sidebar container ─────────────────────────────────────────────────── */
+
+.myPacks {
+    background: #ffffff;
+    border-radius: 12px;
+    padding: 16px;
+    border: 1px solid #eceadf;
+    align-self: start;
+    max-height: calc(100vh - 120px);
+    overflow-x: hidden;
+    overflow-y: auto;
+    min-width: 0;
+    width: 100%;
+    box-sizing: border-box;
+}
+
+/* ── Header row ────────────────────────────────────────────────────────── */
+
+.myPacksHeaderRow {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 8px;
+    margin-bottom: 12px;
+}
+
+.myPacksHeaderRow .myPacksTitle {
+    margin-bottom: 0;
+    flex: 1;
+    min-width: 0;
+}
+
+.myPacksDockToggleBtn {
+    flex-shrink: 0;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    width: 32px;
+    height: 32px;
+    padding: 0;
+    border: 1px solid #e0dbd0;
+    border-radius: 8px;
+    background: #faf8f4;
+    color: #3d3c36;
+    cursor: pointer;
+}
+
+.myPacksDockToggleBtn:hover:not(:disabled) {
+    background: #f0ebe0;
+}
+
+.myPacksDockToggleBtn:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+/* ── Pack row ──────────────────────────────────────────────────────────── */
+
+.myPackRow {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start;
+    gap: 10px;
+    padding: 8px 0;
+    border-bottom: 1px solid #f5f2e3;
+}
+
+.myPackRow > div:first-child {
+    min-width: 0;
+    flex: 1;
+}
+
+.myPackRow:last-child {
+    border-bottom: none;
+}
+
+.myPackRowHighlight {
+    background: rgba(251, 170, 105, 0.12);
+    margin: 0 -4px;
+    padding: 8px 4px;
+    border-radius: 6px;
+}
+
+.myPackRowAside {
+    display: flex;
+    flex-direction: column;
+    gap: 2px;
+    flex-shrink: 0;
+    align-items: center;
+}
+
+/* ── Icon buttons ──────────────────────────────────────────────────────── */
+
+.myPackIconBtn {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    width: 26px;
+    height: 26px;
+    padding: 0;
+    border: none;
+    border-radius: 6px;
+    background: transparent;
+    color: #7a786d;
+    cursor: pointer;
+    font-size: 0.75rem;
+    line-height: 1;
+}
+
+.myPackIconBtn:hover:not(:disabled) {
+    background: #f3f1e6;
+    color: #1e1f23;
+}
+
+.myPackIconBtn:focus-visible {
+    outline: 2px solid rgba(251, 170, 105, 0.85);
+    outline-offset: 1px;
+}
+
+.myPackIconBtn:disabled {
+    opacity: 0.45;
+    cursor: not-allowed;
+}
+
+/* ── Title block ───────────────────────────────────────────────────────── */
+
+.myPackName {
+    font-size: 0.875rem;
+    margin: 0;
+    font-weight: 500;
+    overflow-wrap: anywhere;
+    word-break: break-word;
+}
+
+.myPackTitleBlock {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 6px;
+    margin-bottom: 2px;
+    width: 100%;
+    min-width: 0;
+}
+
+.myPackNameRow {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    width: 100%;
+    flex-wrap: wrap;
+}
+
+.myPackFactoryNameBtn {
+    background: none;
+    border: none;
+    padding: 0;
+    margin: 0;
+    font-size: 0.875rem;
+    font-weight: 500;
+    color: inherit;
+    cursor: pointer;
+    text-align: left;
+    font-family: inherit;
+    overflow-wrap: anywhere;
+    word-break: break-word;
+}
+
+.myPackFactoryNameBtn:hover {
+    color: #fbaa69;
+    text-decoration: underline;
+}
+
+.myPackTitleBlock .myPackName {
+    width: 100%;
+}
+
+.myPackVersion {
+    font-size: 0.75rem;
+    color: #7a786d;
+    overflow-wrap: anywhere;
+    word-break: break-word;
+    display: inline-block;
+    max-width: 100%;
+}
+
+/* ── Fork lineage lines ────────────────────────────────────────────────── */
+
+.hubForkOfLine {
+    font-size: 0.65rem;
+    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace;
+    margin: 4px 0 0;
+    color: #7a786d;
+    overflow-wrap: anywhere;
+    word-break: break-word;
+}
+
+/* ── Fork family group ─────────────────────────────────────────────────── */
+
+.hubForkFamily {
+    padding: 8px 0 14px;
+    margin-bottom: 4px;
+    border-bottom: 1px solid #eceadf;
+    min-width: 0;
+}
+
+.hubForkFamily:last-child {
+    border-bottom: none;
+}
+
+.hubForkFamilyToolbar {
+    margin-bottom: 6px;
+}
+
+.hubForkRailTitleRow {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    gap: 10px;
+    margin-bottom: 4px;
+}
+
+.hubPackFactoryTitleBtn {
+    background: none;
+    border: none;
+    padding: 0;
+    margin: 0;
+    font-size: 0.8125rem;
+    font-weight: 600;
+    color: #1e1f23;
+    cursor: pointer;
+    text-align: left;
+    font-family: inherit;
+}
+
+.hubPackFactoryTitleBtn:hover {
+    color: #fbaa69;
+    text-decoration: underline;
+}
+
+.hubForkFamilyRailTitle {
+    font-size: 0.8125rem;
+    font-weight: 600;
+    margin: 0 0 4px;
+    color: #1e1f23;
+}
+
+.hubForkFamilyRailMeta {
+    font-size: 0.6875rem;
+    font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace;
+    margin: 0;
+    color: #7a786d;
+    overflow-wrap: break-word;
+    word-break: break-word;
+}
+
+.hubForkFamilySelectLabel {
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+    font-size: 0.65rem;
+    text-transform: uppercase;
+    letter-spacing: 0.06em;
+    color: #7a786d;
+    margin-bottom: 8px;
+}
+
+/* ── Empty state ───────────────────────────────────────────────────────── */
+
+.empty {
+    color: #7a786d;
+    text-align: center;
+    padding: 32px 16px;
+    font-size: 0.875rem;
+}
+
